+++ NemesisEuchre.GameEngine/PlayerDecisionEngine/IPlayerActor.cs
@@ -5,9 +5,9 @@ namespace NemesisEuchre.GameEngine.PlayerDecisionEngine;
 
 public interface IPlayerActor
 {
-    CallTrumpDecision CallTrump(List<Card> cardsInHand, Card upCard, RelativePlayerPosition dealerPosition, int teamScore, int opponentScore, CallTrumpDecision[] validCallTrumpDecisions);
+    CallTrumpDecision CallTrump(List<Card> cardsInHand, Card upCard, RelativePlayerPosition dealerPosition, short teamScore, short opponentScore, CallTrumpDecision[] validCallTrumpDecisions);
 
-    RelativeCard DiscardCard(List<RelativeCard> cardsInHand, RelativeDeal? currentDeal, int teamScore, int opponentScore, RelativeCard[] validCardsToDiscard);
+    RelativeCard DiscardCard(List<RelativeCard> cardsInHand, RelativeDeal? currentDeal, short teamScore, short opponentScore, RelativeCard[] validCardsToDiscard);
 
-    RelativeCard PlayCard(List<RelativeCard> cardsInHand, RelativeDeal? currentDeal, int teamScore, int opponentScore, RelativeCard[] validCardsToPlay);
+    RelativeCard PlayCard(List<RelativeCard> cardsInHand, RelativeDeal? currentDeal, short teamScore, short opponentScore, RelativeCard[] validCardsToPlay);
 }
